+++ include/grpcpp/ext/proto_server_reflection_plugin.h
@@ -1,6 +1,6 @@
 /*
  *
- * Copyright 2019 gRPC authors.
+ * Copyright 2015 gRPC authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,17 +19,35 @@
 #ifndef GRPCPP_EXT_PROTO_SERVER_REFLECTION_PLUGIN_H
 #define GRPCPP_EXT_PROTO_SERVER_REFLECTION_PLUGIN_H
 
-#include <grpcpp/ext/proto_server_reflection_plugin_impl.h>
+#include <grpcpp/impl/server_builder_plugin.h>
+#include <grpcpp/support/config.h>
+
+namespace grpc_impl {
+class ServerInitializer;
+}  // namespace grpc_impl
 
 namespace grpc {
+class ProtoServerReflection;
+
 namespace reflection {
 
-typedef ::grpc_impl::reflection::ProtoServerReflectionPlugin
-    ProtoServerReflectionPlugin;
+class ProtoServerReflectionPlugin : public ::grpc::ServerBuilderPlugin {
+ public:
+  ProtoServerReflectionPlugin();
+  ::std::string name() override;
+  void InitServer(::grpc_impl::ServerInitializer* si) override;
+  void Finish(::grpc_impl::ServerInitializer* si) override;
+  void ChangeArguments(const ::std::string& name, void* value) override;
+  bool has_async_methods() const override;
+  bool has_sync_methods() const override;
 
-static inline void InitProtoReflectionServerBuilderPlugin() {
-  ::grpc_impl::reflection::InitProtoReflectionServerBuilderPlugin();
-}
+ private:
+  std::shared_ptr<grpc::ProtoServerReflection> reflection_service_;
+};
+
+/// Add proto reflection plugin to \a ServerBuilder.
+/// This function should be called at the static initialization time.
+void InitProtoReflectionServerBuilderPlugin();
 
 }  // namespace reflection
 }  // namespace grpc
