+++ src/cpp/ext/proto_server_reflection_plugin.cc
@@ -23,7 +23,7 @@
 
 #include "src/cpp/ext/proto_server_reflection.h"
 
-namespace grpc_impl {
+namespace grpc {
 namespace reflection {
 
 ProtoServerReflectionPlugin::ProtoServerReflectionPlugin()
@@ -80,4 +80,4 @@ struct StaticProtoReflectionPluginInitializer {
 } static_proto_reflection_plugin_initializer;
 
 }  // namespace reflection
-}  // namespace grpc_impl
+}  // namespace grpc
